+++ android/mwcore/src/main/java/org/smartregister/fhircore/mwcore/ui/patient/details/QuestPatientDetailActivity.kt
@@ -26,10 +26,10 @@ import org.smartregister.fhircore.engine.ui.base.BaseMultiLanguageActivity
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireConfig
 import org.smartregister.fhircore.engine.ui.theme.AppTheme
-import org.smartregister.fhircore.quest.QuestApplication
-import org.smartregister.fhircore.quest.R
-import org.smartregister.fhircore.quest.data.patient.PatientRepository
-import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
+import org.smartregister.fhircore.mwcore.MwCoreApplication
+import org.smartregister.fhircore.mwcore.R
+import org.smartregister.fhircore.mwcore.data.patient.PatientRepository
+import org.smartregister.fhircore.mwcore.ui.patient.register.PatientItemMapper
 import timber.log.Timber
 
 class QuestPatientDetailActivity : BaseMultiLanguageActivity() {
@@ -40,10 +40,10 @@ class QuestPatientDetailActivity : BaseMultiLanguageActivity() {
     super.onCreate(savedInstanceState)
 
     patientId = intent.extras?.getString(QuestionnaireActivity.QUESTIONNAIRE_ARG_PATIENT_KEY) ?: "1"
-    val fhirEngine = (QuestApplication.getContext() as ConfigurableApplication).fhirEngine
+    val fhirEngine = (MwCoreApplication.getContext() as ConfigurableApplication).fhirEngine
     val repository = PatientRepository(fhirEngine, PatientItemMapper)
     val viewModel =
-      QuestPatientDetailViewModel.get(this, application as QuestApplication, repository, patientId)
+      QuestPatientDetailViewModel.get(this, application as MwCoreApplication, repository, patientId)
 
     viewModel.setOnBackPressListener(this::onBackPressListener)
     viewModel.setOnMenuItemClickListener(this::onMenuItemClickListener)
