+++ android/mwcore/src/main/java/org/smartregister/fhircore/mwcore/ui/patient/details/QuestPatientDetailScreen.kt
@@ -81,7 +81,7 @@ import org.smartregister.fhircore.engine.util.extension.asDdMmmYyyy
 import org.smartregister.fhircore.engine.util.extension.extractAge
 import org.smartregister.fhircore.engine.util.extension.extractGender
 import org.smartregister.fhircore.engine.util.extension.extractName
-import org.smartregister.fhircore.quest.R
+import org.smartregister.fhircore.mwcore.R
 
 const val TOOLBAR_TITLE = "toolbarTitle"
 const val TOOLBAR_BACK_ARROW = "toolbarBackArrow"
