+++ android/mwcore/src/main/java/org/smartregister/fhircore/mwcore/ui/patient/details/QuestPatientTestResultActivity.kt
@@ -22,9 +22,9 @@ import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplicati
 import org.smartregister.fhircore.engine.ui.base.BaseMultiLanguageActivity
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity
 import org.smartregister.fhircore.engine.ui.theme.AppTheme
-import org.smartregister.fhircore.quest.QuestApplication
-import org.smartregister.fhircore.quest.data.patient.PatientRepository
-import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
+import org.smartregister.fhircore.mwcore.MwCoreApplication
+import org.smartregister.fhircore.mwcore.data.patient.PatientRepository
+import org.smartregister.fhircore.mwcore.ui.patient.register.PatientItemMapper
 
 class QuestPatientTestResultActivity : BaseMultiLanguageActivity() {
 
@@ -34,10 +34,10 @@ class QuestPatientTestResultActivity : BaseMultiLanguageActivity() {
     super.onCreate(savedInstanceState)
 
     patientId = intent.extras?.getString(QuestionnaireActivity.QUESTIONNAIRE_ARG_PATIENT_KEY) ?: "1"
-    val fhirEngine = (QuestApplication.getContext() as ConfigurableApplication).fhirEngine
+    val fhirEngine = (MwCoreApplication.getContext() as ConfigurableApplication).fhirEngine
     val repository = PatientRepository(fhirEngine, PatientItemMapper)
     val viewModel =
-      QuestPatientDetailViewModel.get(this, application as QuestApplication, repository, patientId)
+      QuestPatientDetailViewModel.get(this, application as MwCoreApplication, repository, patientId)
 
     viewModel.setOnBackPressListener(this::onBackPressListener)
